adding cleanup to the test suite.
this program will kill any remeaining instances running.  Anything that
was running before the tests were run will remain.  However if instnaces
are started out of band but associated with this account, they will be
killed.  So the test suite should only be run by an account that is not
starting other instances in parallel.